+++ NEWS.md
@@ -6,6 +6,8 @@ actual function in which the error occurred.
 * The DESCRIPTION file in jtools 2.2.0 wrongly defined the minimum version of 
 R to use this package as R "3.50." It should have been R 3.5.0 and now reflects
 that it should work on versions of R greater than or equal to 3.5.0. (#126)
+* The `effect_plot()` function now accepts the argument `line.colors` to allow
+users to request different colors for the line/interval and plotted points. 
 
 # jtools 2.2.0